Day11 (#11)
* Part One

* Part Two

Edit: the only thing worryLevel is used for is to determine where to
send the item. The actual value is irrelevant so long as the monkey
test holds the same result. So therefore we can modulo using the common
divisor of all the monkey test values and use the remainder as the worry
level.

Not really understanding why this is needed or why it works

https://github.com/BenjaminB64/aoc2022/blob/main/11/monkey/models.go\#L106

Shout outs to that man

Tests broken but I am not going to fix them now